rename: change package in user, login class
